Working on making a tutorial, messed something up somehow, retrieving code from previous version
