+++ Assets/Tutorial.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: ebefccd8db5fe9b48b3c192996a5af7f
+folderAsset: yes
+timeCreated: 1510788554
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
